**How It Works**
To fully explain this we need to step back a bit first. Let’s assume that you are in a
fresh Lisp image with the package CL-USER being your current package. You want
to define a function using the well-known macro DEFUN, and thus you need to refer
to the macro’s name, which is a symbol. There are several ways to do this: